+++ android/app/src/main/kotlin/com/example/sizzle/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.sizzle
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()